pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["nurbol0tt <fairtail764@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-yasg = "^1.21.7"
django-jet-reboot = "^1.3.3"
django-split-settings = "^1.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
django-cors-headers = "^4.2.0"
six = "^1.16.0"
djangorestframework-simplejwt = "^5.3.0"
gunicorn = "^21.2.0"
celery = {extras = ["redis"], version = "^5.3.1"}
django-debug-toolbar = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
